+++ router/producer_router.go
@@ -5,6 +5,9 @@ import (
 	"net/http"
 	"net/http/httputil"
 	"net/url"
+
+	"github.com/BaritoLog/barito-router/appcontext"
+	"github.com/BaritoLog/barito-router/instrumentation"
 )
 
 const (
@@ -12,6 +15,8 @@ const (
 	AppGroupSecretHeaderName = "X-App-Group-Secret"
 	AppNameHeaderName        = "X-App-Name"
 	KeyProducer              = "producer"
+	AppNoProfilePath         = "api/producer_no_profile"
+	AppNoSecretPath          = "api/no_secret"
 )
 
 type ProducerRouter interface {
@@ -26,15 +31,17 @@ type producerRouter struct {
 	profileByAppGroupPath string
 
 	client *http.Client
+	appCtx *appcontext.AppContext
 }
 
-func NewProducerRouter(addr, marketUrl, profilePath string, profileByAppGroupPath string) ProducerRouter {
+func NewProducerRouter(addr, marketUrl, profilePath string, profileByAppGroupPath string, appCtx *appcontext.AppContext) ProducerRouter {
 	return &producerRouter{
 		addr:                  addr,
 		marketUrl:             marketUrl,
 		profilePath:           profilePath,
 		profileByAppGroupPath: profileByAppGroupPath,
 		client:                createClient(),
+		appCtx:                appCtx,
 	}
 }
 
@@ -61,12 +68,19 @@ func (p *producerRouter) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 	if appSecret == "" {
 		if appGroupSecret != "" && appName != "" {
 			profile, err = fetchProfileByAppGroupSecret(p.client, p.marketUrl, p.profileByAppGroupPath, appGroupSecret, appName)
+			if profile != nil {
+				instrumentation.RunTransaction(p.appCtx.NewRelicApp(), p.profileByAppGroupPath, w, req)
+			}
 		} else {
 			onNoSecret(w)
+			instrumentation.RunTransaction(p.appCtx.NewRelicApp(), AppNoSecretPath, w, req)
 			return
 		}
 	} else {
 		profile, err = fetchProfileByAppSecret(p.client, p.marketUrl, p.profilePath, appSecret)
+		if profile != nil {
+			instrumentation.RunTransaction(p.appCtx.NewRelicApp(), p.profilePath, w, req)
+		}
 	}
 	if err != nil {
 		onTradeError(w, err)
@@ -75,6 +89,8 @@ func (p *producerRouter) ServeHTTP(w http.ResponseWriter, req *http.Request) {
 
 	if profile == nil {
 		onNoProfile(w)
+		instrumentation.RunTransaction(p.appCtx.NewRelicApp(), AppNoProfilePath, w, req)
+
 		return
 	}