+++ router/producer_router_test.go
@@ -9,6 +9,8 @@ import (
 	"github.com/BaritoLog/go-boilerplate/httpkit"
 	. "github.com/BaritoLog/go-boilerplate/testkit"
 	"github.com/hashicorp/consul/api"
+	"github.com/newrelic/go-agent"
+	"github.com/BaritoLog/barito-router/appcontext"
 )
 
 func TestProducerRouter_Ping(t *testing.T) {
@@ -17,7 +19,11 @@ func TestProducerRouter_Ping(t *testing.T) {
 
 	req, _ := http.NewRequest("GET", "/ping", nil)
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusOK)
@@ -28,14 +34,22 @@ func TestProducerRouter_FetchError(t *testing.T) {
 	req, _ := http.NewRequest("GET", "/", nil)
 	req.Header.Add("X-App-Secret", "some-secret")
 
-	router := NewProducerRouter(":65500", "http://wrong-market", "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":65500", "http://wrong-market", "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusBadGateway)
 }
 
 func TestProducerRouter_NoSecret(t *testing.T) {
-	router := NewProducerRouter(":65500", "http://wrong-market", "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":65500", "http://wrong-market", "profilePath", "profileByAppGroupPath", appCtx)
 
 	req, _ := http.NewRequest("GET", "/", nil)
 	resp := RecordResponse(router.ServeHTTP, req)
@@ -50,7 +64,11 @@ func TestProducerRouter_NoProfile(t *testing.T) {
 	req, _ := http.NewRequest("GET", "/", nil)
 	req.Header.Add("X-App-Secret", "some-secret")
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusNotFound)
@@ -64,7 +82,11 @@ func TestProducerRouter_WithAppGroupSecret_NoProfile(t *testing.T) {
 	req.Header.Add("X-App-Group-Secret", "some-secret")
 	req.Header.Add("X-App-Name", "some-name")
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profileByAppGroupPath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profileByAppGroupPath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusNotFound)
@@ -79,7 +101,11 @@ func TestProducerRouter_ConsulError(t *testing.T) {
 	req, _ := http.NewRequest("GET", "/", nil)
 	req.Header.Add("X-App-Secret", "some-secret")
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusFailedDependency)
@@ -109,7 +135,11 @@ func TestProducerRouter_WithAppSecret(t *testing.T) {
 	req, _ := http.NewRequest(http.MethodGet, "http://localhost", bytes.NewBuffer(testPayload))
 	req.Header.Add("X-App-Secret", "some-secret")
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 	FatalIfWrongResponseStatus(t, resp, http.StatusTeapot)
 	FatalIfWrongResponseBody(t, resp, "some-target")
@@ -140,7 +170,11 @@ func TestProducerRouter_WithAppGroupSecret(t *testing.T) {
 	req.Header.Add("X-App-Group-Secret", "some-secret")
 	req.Header.Add("X-App-Name", "some-name")
 
-	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath")
+	config := newrelic.NewConfig("barito-router", "")
+	config.Enabled = false
+	appCtx := appcontext.NewAppContext(config)
+
+	router := NewProducerRouter(":45500", marketServer.URL, "profilePath", "profileByAppGroupPath", appCtx)
 	resp := RecordResponse(router.ServeHTTP, req)
 
 	FatalIfWrongResponseStatus(t, resp, http.StatusTeapot)