+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+**0.5.4**
+
+- add new param: barito market access token and send it to market when calling fetch profile by cluster name
+- use v2 API when fetching profile by cluster name from market
+
+**0.5.3**
+
+- Add support for newrelic
+
 **0.5.2**
 
 - Log unsuccessful request and return the error message back to the requester